requirements.txt
tensorflow-cpu>=2.0.3
numpy
opencv-python>=4.1.1.26
lxml
tqdm
flask
seaborn
pillow